Move api response to dto
