+++ labellab_mobile/lib/data/remote/dto/register_response.dart
@@ -1,4 +1,3 @@
-
 class RegisterResponse {
   String msg;
   String errField;
