+++ labellab_mobile/lib/data/remote/labellab_api.dart
@@ -3,7 +3,7 @@ import 'dart:io';
 import 'package:labellab_mobile/data/remote/dto/google_user_request.dart';
 import 'package:labellab_mobile/data/remote/dto/login_response.dart';
 import 'package:labellab_mobile/data/remote/dto/register_response.dart';
-import 'package:labellab_mobile/model/api_response.dart';
+import 'package:labellab_mobile/data/remote/dto/api_response.dart';
 import 'package:labellab_mobile/model/auth_user.dart';
 import 'package:labellab_mobile/model/classification.dart';
 import 'package:labellab_mobile/model/group.dart';
